requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.23.2
pillow==10.1.0
torch==2.1.2
transformers==4.36.2
python-multipart==0.0.6
numpy==1.26.2
httpx==0.28.1
